langchain/product-assisstant/requirements.txt
pydantic
python-dotenv
langchain
langchain-openai
langchain-groq
langchain-core
langchain-community
flask
chromadb
jq
streamlit